requirements.txt
PyMuPDF==1.27.2
langchain-openai==1.1.11
langgraph==1.1.2
pydantic==2.12.5
python-dotenv==1.2.2
pytest==9.0.2
